Enable Wake On LAN server functionality
Added a "mac_address" element to devices in the config.json file. If set,
a "Wake" button will be displayed when the device is marked offline.
Added "requirements.txt" which can be used to help ensure pip
dependancies are installed (currently just "wakeonlan").
Updated README to include new functionality and installation steps.